+++ src/main/java/net/silentchaos512/lib/data/client/package-info.java
@@ -1,0 +1,4 @@
+@NullMarked
+package net.silentchaos512.lib.data.client;
+
+import org.jspecify.annotations.NullMarked;