+++ src/main/java/net/silentchaos512/lib/event/InitialSpawnItems.java
@@ -1,5 +1,6 @@
 package net.silentchaos512.lib.event;
 
+import com.mojang.datafixers.types.Func;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.resources.Identifier;
 import net.minecraft.world.entity.player.Player;
@@ -18,8 +19,8 @@ import java.util.function.Function;
 import java.util.function.Supplier;
 
 /**
- * Can be used to give players items when they first join a world. Call {@link
- * #add(net.minecraft.resources.Identifier, Supplier)} in either init or post-init. <em>This should be used
+ * Can be used to give players items when they first join a world. Call {@link #addItem(Identifier, Function)} or
+ * {@link #addItems(Identifier, Function)} in either init or post-init. <em>This should be used
  * sparingly</em>; we spawn with enough junk already. It is recommended to have config options to
  * disable spawn items.
  *
@@ -42,44 +43,35 @@ public final class InitialSpawnItems {
      * is given. If the item is given, the player will never receive a spawn item with the same
      * {@code key} again. If no item is given, the key will not be marked as given.
      *
-     * @param key   The key to uniquely identify the spawn item
-     * @param stack The {@link ItemStack} supplier. If it returns an empty stack, nothing is given.
-     * @deprecated Use {@link #add(Identifier, Function)} instead
-     */
-    @Deprecated
-    public static void add(Identifier key, Supplier<ItemStack> stack) {
-        INSTANCE.spawnItems.put(key, p -> {
-            ItemStack s = stack.get();
-            return s.isEmpty() ? Collections.emptyList() : Collections.singleton(s);
-        });
-    }
-
-    /**
-     * Add a spawn item. If the supplier returns an empty stack ({@link ItemStack#EMPTY}), nothing
-     * is given. If the item is given, the player will never receive a spawn item with the same
-     * {@code key} again. If no item is given, the key will not be marked as given.
-     *
      * @param key         The key to uniquely identify the spawn item
      * @param itemFactory The item stack producer. Should not contain empty stacks.
+     * @deprecated Use {@link #addItem(Identifier, Function)} or {@link #addItems(Identifier, Function)} instead
      */
+    @Deprecated
     public static void add(Identifier key, Function<Player, Collection<ItemStack>> itemFactory) {
-        INSTANCE.spawnItems.put(key, itemFactory);
+        addItems(key, itemFactory);
+    }
+
+    public static void addItem(Identifier key, Function<Player, ItemStack> itemSupplier) {
+        INSTANCE.spawnItems.put(key, p -> Collections.singleton(itemSupplier.apply(p)));
+    }
+
+    public static void addItems(Identifier key, Function<Player, Collection<ItemStack>> itemsSupplier) {
+        INSTANCE.spawnItems.put(key, itemsSupplier);
     }
 
     private void onPlayerLoggedIn(PlayerEvent.PlayerLoggedInEvent event) {
         Player player = event.getEntity();
         CompoundTag givenItems = PlayerUtils.getPersistedDataSubcompound(player, NBT_KEY);
 
-        spawnItems.forEach((key, factory) -> handleSpawnItems(player, givenItems, key, factory.apply(player)));
+        spawnItems.forEach((key, factory) -> handleSpawnItems(player, givenItems, key, factory));
     }
 
-    private static void handleSpawnItems(Player player, CompoundTag givenItems, Identifier key, Collection<ItemStack> items) {
-        if (items.isEmpty()) return;
-
+    private static void handleSpawnItems(Player player, CompoundTag givenItems, Identifier key, Function<Player, Collection<ItemStack>> itemSupplier) {
         String nbtKey = key.toString().replace(':', '.');
         var hasBeenGiven = givenItems.getBoolean(nbtKey);
         if (hasBeenGiven.isEmpty() || !hasBeenGiven.get()) {
-            items.forEach(stack -> {
+            itemSupplier.apply(player).forEach(stack -> {
                 SilentLib.LOGGER.debug("Giving player {} spawn item \"{}\": {}", player.getScoreboardName(), nbtKey, stack);
                 PlayerUtils.giveItem(player, stack);
                 givenItems.putBoolean(nbtKey, true);